refactor(#78): 로비 맵 접근 정책 분리
